musinsa/requirements.txt
python>=3.10
opencv-python
opencv-contrib-python
numpy
Pillow
pyyaml
scipy
matplotlib
pandas
faiss-cpu
torch
torchvision
torchaudio
transformers
huggingface-hub
sentencepiece
ultralytics
segment-anything-hq
groundingdino-py
diffusers[torch]
accelerate
compel
rembg
